requirements-freeze.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-ws==0.7.2
idna==3.10
inflection==0.5.1
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lmstudio==1.4.1
msgspec==0.19.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.99.1
openapi==2.0.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyExifTool==0.5.6
pyparsing==3.2.3
pytesseract==0.3.13
python-dotenv==1.1.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.27.0
rsa==4.9.1
sniffio==1.3.1
tiktoken==0.10.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
wsproto==1.2.0
